Fully migrate the coupon field up to the checkout main content component.
